Allow expiry of all logs
Before this change, the copy lambda creates a log group which never
expires. When you manage lots of AWS accounts and you want to make sure
that all cloudwatch log groups expire at some point, this can get very
annoying.

I renamed the logical id of the function, so that a new function is
created and therefore we get a chance to create the log group. Otherwise
the stack would fail since there would be already an existing log group.