Add types/jest in package, not release
Add the types/jest module in the package.json as a dev dependency
instead of just adding it as a dev dependency during the release.
The build failed in the release process because of this toThrowError
vs toThrow, this should move those failures to development time instead
of when we go to release.

Error:https://gitlab.ddbuild.io/DataDog/datadog-cloudformation-macro/-/jobs/983337668#L124