Use JWT to get GH Token (#201)
Previously I thought that we could use the JWT to as the token for github, but
we actually need to use that to get the github token, and then everything else
should work the same.  I tried this by checking the auth status on a test run,
and it seems to work.